+++ src/ansys/additive/core/additive.py
@@ -456,11 +456,11 @@ class Additive:
             raise ValueError("A material is not assigned to the simulation input")
 
         if (
-            isinstance(simulation_input, Microstructure3DInput)
+            isinstance(simulation_input, (Microstructure3DInput, ThermalHistoryInput))
             and self.enable_beta_features is False
         ):
             raise BetaFeatureNotEnabledError(
-                "3D microstructure simulations require beta features to be enabled.\n"
+                "This simulation requires beta features to be enabled.\n"
                 "Set enable_beta_features=True when creating the Additive client."
             )